python/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
Pillow>=10.0.0
pydantic>=2.0.0
typing-extensions>=4.7.0 